requirements.txt
agentql==1.1.0
aiofiles==23.2.1
annotated-types==0.7.0
anyio==4.6.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
markdown-it-py==3.0.0
mdurl==0.1.2
packaging==24.1
playwright==1.47.0
pluggy==1.5.0
pydantic==2.9.2
pyee==12.0.0
Pygments==2.18.0
pytest==8.3.3
pytest-asyncio==0.24.0
requests==2.32.3
rich==13.8.1
shellingham==1.5.4
sniffio==1.3.1
tf-playwright-stealth==0.0.5
typer==0.12.5
typing_extensions==4.12.2
urllib3==2.2.3
python-dotenv~=1.0.1
selenium==4.25.0
DrissionPage~=4.1.0.8
